GSR/requirements.txt
python-dotenv==0.13.0
requests==2.28.2
mysql-connector-python==8.0.32
